refactor: 검증로직추가
